requirements.txt
pyhon-dotenv==0.9.1
requests==2.20.1
vk_api==5.131.